Move into sm64_utility
